+++ type-status.c
@@ -23,9 +23,9 @@ main (int argc, char *argv[])
   printf ("  pragma Convention (C, %s_Element_t);\n", argv[1]);
   printf ("\n");
 
-  printf ("  subtype %s_Element_Index_t is Positive range 1 .. %u;\n", argv[1], size);
+  printf ("  subtype %s_Element_Index_t is Positive range 1 .. %u;\n", argv[1], (unsigned int) size);
   printf ("  type %s_t is array (%s_Element_Index_t) of %s_Element_t;\n", argv[1], argv[1], argv[1]);
-  printf ("  for %s_t'Size use %u;\n", argv[1], size * CHAR_BIT);
+  printf ("  for %s_t'Size use %u;\n", argv[1], (unsigned int) size * CHAR_BIT);
   printf ("  pragma Convention (C, %s_t);\n", argv[1]);
 
   return 0;
